Add api gateway description in the swagger UI
