+++ src/main/java/org/semantics/apigateway/APIGatewayApplication.java
@@ -1,0 +1,44 @@
+package org.semantics.apigateway;
+
+import io.swagger.v3.oas.annotations.OpenAPIDefinition;
+import io.swagger.v3.oas.annotations.info.Info;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+
+import java.util.concurrent.Executor;
+
+@SpringBootApplication
+@EnableConfigurationProperties
+@EnableAsync
+@OpenAPIDefinition(
+		info = @Info(title = "API Gateway Documentation", version = "1.0", description = "The TS4NFDI Federated Service is an advanced, dynamic solution designed to perform federated calls across multiple Terminology Services (TS) within NFDI. It is particularly tailored for environments where integration and aggregation of diverse data sources are essential. The service offers search capabilities, enabling users to refine search results based on specific criteria, and supports responses in both JSON and JSON-LD formats.\n" +
+				"\n" +
+				"A standout feature of this service is its dynamic nature, governed by a JSON configuration file. This design choice allows for easy extension and customization of the service to include new TS or modify existing configurations."),
+		tags = {
+				@Tag(name = "Search" ),
+				@Tag(name = "Artefacts"),
+		}
+)
+public class APIGatewayApplication {
+
+	public static void main(String[] args) {
+
+		SpringApplication.run(APIGatewayApplication.class, args);
+	}
+
+	@Bean
+	public Executor taskExecutor() {
+		ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+		executor.setCorePoolSize(10);
+		executor.setMaxPoolSize(10);
+		executor.setQueueCapacity(500);
+		executor.setThreadNamePrefix("Async-");
+		executor.initialize();
+		return executor;
+	}
+}
