--- src/main/java/org/semantics/apigateway/FederatedSearchApplication.java
@@ -1,33 +1,0 @@
-package org.semantics.apigateway;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-
-import java.util.concurrent.Executor;
-
-@SpringBootApplication
-@EnableConfigurationProperties
-@EnableAsync
-public class FederatedSearchApplication {
-
-	public static void main(String[] args) {
-
-		SpringApplication.run(FederatedSearchApplication.class, args);
-	}
-
-	@Bean
-	public Executor taskExecutor() {
-		ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-		executor.setCorePoolSize(10);
-		executor.setMaxPoolSize(10);
-		executor.setQueueCapacity(500);
-		executor.setThreadNamePrefix("Async-");
-		executor.initialize();
-		return executor;
-	}
-}
